chore: undo Response->SimpleResponse rename
