+++ python/examples/calls/custom_client_google.py
@@ -11,6 +11,6 @@ def recommend_book(genre: str):
     return f"Recommend a {genre} book"
 
 
-response: llm.Response = recommend_book("fantasy")
+response: llm.BaseResponse = recommend_book("fantasy")
 print(response.text)
 # "Here are a few highly recommended fantasy books..."
