+++ python/mirascope/llm/__init__.py
@@ -57,9 +57,9 @@ from .response_formatting.decorator import response_format
 from .responses import (
     AsyncStream,
     AsyncStructuredStream,
+    BaseResponse,
     ContextResponse,
     Response,
-    SimpleResponse,
     Stream,
     StreamChunk,
     StructuredStream,
@@ -95,10 +95,10 @@ __all__ = [
     "PermissionError",
     "Prompt",
     "RateLimitError",
-    "Response",
+    "BaseResponse",
     "ResponseFormat",
     "ServerError",
-    "SimpleResponse",
+    "Response",
     "Stream",
     "StreamChunk",
     "StructuredAgent",
