+++ python/mirascope/llm/responses/__init__.py
@@ -4,6 +4,7 @@ from .async_context_stream import AsyncContextStream
 from .async_context_structured_stream import AsyncContextStructuredStream
 from .async_stream import AsyncStream
 from .async_structured_stream import AsyncStructuredStream
+from .base_response import BaseResponse
 from .content import (
     BaseResponseContent,
     ContextResponseContent,
@@ -16,7 +17,6 @@ from .context_stream_chunk import ContextStreamChunk
 from .context_structured_stream import ContextStructuredStream
 from .finish_reason import FinishReason
 from .response import Response
-from .simple_response import SimpleResponse
 from .stream import Stream
 from .stream_chunk import StreamChunk
 from .structured_stream import StructuredStream
@@ -34,9 +34,9 @@ __all__ = [
     "ContextStreamChunk",
     "ContextStructuredStream",
     "FinishReason",
-    "Response",
+    "BaseResponse",
     "ResponseContent",
-    "SimpleResponse",
+    "Response",
     "SimpleResponseContent",
     "Stream",
     "StreamChunk",
