+++ python/mirascope/llm/responses/context_response.py
@@ -13,12 +13,15 @@ from typing_extensions import TypeVar
 from ..tools import ContextTool
 from ..types import Jsonable
 from .content import ContextResponseContent
-from .response import Response
+from .base_response import BaseResponse
 
 DepsT = TypeVar("DepsT", default=None)
 T = TypeVar("T", bound=object | None, default=None)
 
 
 @dataclass
-class ContextResponse(Response[ContextResponseContent, ContextTool[..., Jsonable, DepsT], T], Generic[DepsT, T]):
+class ContextResponse(
+    BaseResponse[ContextResponseContent, ContextTool[..., Jsonable, DepsT], T],
+    Generic[DepsT, T],
+):
     """The response generated by an LLM."""
