--- python/mirascope/llm/responses/simple_response.py
@@ -1,21 +1,0 @@
-"""Interfaces for LLM responses.
-
-This module defines interfaces for the responses returned by language models,
-including methods for formatting the response according to a specified format.
-"""
-
-from dataclasses import dataclass
-
-from typing_extensions import TypeVar
-
-from ..tools import Tool
-from ..types import Jsonable
-from .content import SimpleResponseContent
-from .response import Response
-
-T = TypeVar("T", bound=object | None, default=None)
-
-
-@dataclass
-class SimpleResponse(Response[SimpleResponseContent, Tool[..., Jsonable], T]):
-    """The response generated by an LLM."""
